Next step towards a robust indexing of mails. Including some minor fixes and the possibility of using job priorities.

+++ com.openexchange.index.solr/src/com/openexchange/index/solr/internal/attachments/SolrAttachmentIndexAccess.java
@@ -184,7 +184,9 @@ public class SolrAttachmentIndexAccess extends AbstractSolrIndexAccess<Attachmen
         }
         
         String deleteQuery = buildQueryStringWithOr(SolrAttachmentField.UUID.solrName(), uuids);
-        deleteDocumentsByQuery(deleteQuery);
+        if (deleteQuery != null) {
+            deleteDocumentsByQuery(deleteQuery);
+        }
     }
 
     @Override
